Add widgets
Currently only three sizes, as the beta SDK doesn't allow for all the
families outlined in the documentation.

Also offsets don't work, so all the stickies exist in a big stack at the
center of the widget (read: all but one is obscured). I'm going to look
into restricted window sizes as a potential fix, so that offsets never
get set to inappropriate values.